feat: Move component_factory to core/factories/
