+++ src/fapilog/container.py
@@ -13,12 +13,12 @@ from typing import TYPE_CHECKING, Any, Generator, List, Optional
 import structlog
 
 from ._internal.async_lock_manager import ProcessorLockManager
-from ._internal.component_factory import ComponentFactory
 from ._internal.component_registry import ComponentRegistry
 from ._internal.container_logger_factory import ContainerLoggerFactory
 from ._internal.metrics import MetricsCollector
 from ._internal.processor_metrics import ProcessorMetrics
 from ._internal.queue_worker import QueueWorker
+from .core.factories.component_factory import ComponentFactory
 from .core.managers.configuration_manager import ConfigurationManager
 from .core.managers.lifecycle_manager import LifecycleManager
 from .core.managers.metrics_manager import MetricsManager
