+++ src/fapilog/core/factories/component_factory.py
@@ -7,24 +7,24 @@ dependencies.
 
 from typing import TYPE_CHECKING, NamedTuple, Optional
 
-from fapilog._internal.async_lock_manager import ProcessorLockManager
-from fapilog._internal.metrics import MetricsCollector
-from fapilog._internal.processor_metrics import ProcessorMetrics
-from fapilog.core.managers.configuration_manager import ConfigurationManager
-from fapilog.core.managers.lifecycle_manager import LifecycleManager
-from fapilog.core.managers.middleware_manager import MiddlewareManager
-from fapilog.core.managers.sink_manager import SinkManager
-from fapilog.enrichers import (
+from ..._internal.async_lock_manager import ProcessorLockManager
+from ..._internal.metrics import MetricsCollector
+from ..._internal.processor_metrics import ProcessorMetrics
+from ...enrichers import (
     AsyncSmartCache,
     EnricherErrorHandler,
     EnricherHealthMonitor,
     RetryCoordinator,
 )
-from fapilog.monitoring import PrometheusExporter
-from fapilog.settings import LoggingSettings
+from ...monitoring import PrometheusExporter
+from ...settings import LoggingSettings
+from ..managers.configuration_manager import ConfigurationManager
+from ..managers.lifecycle_manager import LifecycleManager
+from ..managers.middleware_manager import MiddlewareManager
+from ..managers.sink_manager import SinkManager
 
 if TYPE_CHECKING:
-    from fapilog.container import LoggingContainer
+    from ...container import LoggingContainer
 
 
 class ManagerSet(NamedTuple):
