+++ tests/test_component_factory.py
@@ -5,10 +5,10 @@ from unittest.mock import Mock
 import pytest
 
 from fapilog._internal.async_lock_manager import ProcessorLockManager
-from fapilog._internal.component_factory import ComponentFactory, ManagerSet
 from fapilog._internal.metrics import MetricsCollector
 from fapilog._internal.processor_metrics import ProcessorMetrics
 from fapilog.container import LoggingContainer
+from fapilog.core.factories.component_factory import ComponentFactory, ManagerSet
 from fapilog.core.managers.configuration_manager import ConfigurationManager
 from fapilog.core.managers.lifecycle_manager import LifecycleManager
 from fapilog.core.managers.middleware_manager import MiddlewareManager
